[KOGITO-4714] - Introducing Quarkus Kogito Processes Extension
