--- kogito-quarkus-parent/kogito-quarkus-extension/kogito-quarkus-deployment/src/main/java/org/kie/kogito/quarkus/deployment/ProcessesCompilationProvider.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito.quarkus.deployment;
-
-import java.util.HashSet;
-import java.util.Set;
-
-import org.kie.kogito.quarkus.common.deployment.KogitoCompilationProvider;
-
-import static java.util.Arrays.asList;
-
-public class ProcessesCompilationProvider extends KogitoCompilationProvider {
-
-    @Override
-    public Set<String> handledExtensions() {
-        return new HashSet<>(asList(".bpmn", ".bpmn2"));
-    }
-}
