--- kogito-quarkus-parent/kogito-quarkus-extension/kogito-quarkus-deployment/src/main/resources/META-INF/services/io.quarkus.deployment.dev.CompilationProvider
@@ -1,1 +1,0 @@
-org.kie.kogito.quarkus.deployment.ProcessesCompilationProvider
