+++ kogito-quarkus-parent/kogito-quarkus-extension/pom.xml
@@ -17,8 +17,6 @@
     <modules>
         <module>kogito-quarkus-deployment</module>
         <module>kogito-quarkus</module>
-        <module>kogito-quarkus-integration-test</module>
-        <module>kogito-quarkus-integration-test-hot-reload</module>
         <module>kogito-quarkus-integration-test-maven-devmode</module>
     </modules>
