+++ kogito-quarkus-parent/kogito-quarkus-processes-extension/kogito-quarkus-processes-deployment/src/main/resources/META-INF/services/io.quarkus.deployment.dev.CompilationProvider
@@ -1,0 +1,1 @@
+org.kie.kogito.quarkus.processes.deployment.ProcessesCompilationProvider
