+++ kogito-quarkus-parent/kogito-quarkus-processes-extension/kogito-quarkus-processes-deployment/src/test/java/org/kie/kogito/quarkus/NotEmptyConstructor.java
