+++ kogito-quarkus-parent/kogito-quarkus-processes-extension/kogito-quarkus-processes-integration-test-hot-reload/src/test/java/io/quarkus/it/kogito/jbpm/JbpmHotReloadTestHelper.java
