+++ kogito-quarkus-parent/kogito-quarkus-processes-extension/kogito-quarkus-processes-integration-test-hot-reload/src/test/resources/orderItems.txt
