+++ kogito-quarkus-parent/kogito-quarkus-processes-extension/kogito-quarkus-processes-integration-test/src/main/java/io/quarkus/it/kogito/jbpm/CalculationService.java
