+++ kogito-quarkus-parent/kogito-quarkus-processes-extension/kogito-quarkus-processes-integration-test/src/main/resources/org/jbpm/orderItems.bpmn
