+++ kogito-quarkus-parent/kogito-quarkus-processes-extension/kogito-quarkus-processes-integration-test/src/test/java/org/kie/kogito/quarkus/jbpm/OrdersProcessIT.java
