+++ kogito-quarkus-parent/kogito-quarkus-processes-extension/pom.xml
@@ -9,14 +9,14 @@
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>kogito-quarkus-serverless-workflow-extension</artifactId>
-  <name>Kogito :: Quarkus Serverless Workflow Extension</name>
+  <artifactId>kogito-quarkus-processes-extension</artifactId>
+  <name>Kogito :: Quarkus Processes Extension</name>
+  <modules>
+    <module>kogito-quarkus-processes</module>
+    <module>kogito-quarkus-processes-deployment</module>
+    <module>kogito-quarkus-processes-integration-test</module>
+    <module>kogito-quarkus-processes-integration-test-hot-reload</module>
+  </modules>
 
   <packaging>pom</packaging>
-
-  <modules>
-    <module>kogito-quarkus-serverless-workflow</module>
-    <module>kogito-quarkus-serverless-workflow-deployment</module>
-    <module>kogito-quarkus-serverless-workflow-integration-test</module>
-  </modules>
 </project>
