+++ kogito-quarkus-parent/kogito-quarkus-serverless-workflow-extension/kogito-quarkus-serverless-workflow-deployment/pom.xml
@@ -24,6 +24,11 @@
       <artifactId>kogito-quarkus-common-deployment</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>kogito-quarkus-processes-deployment</artifactId>
+      <version>${project.version}</version>
+    </dependency>
     <!-- Quarkus Deployment Extensions Dependencies -->
     <dependency>
       <groupId>io.quarkus</groupId>
@@ -46,10 +51,6 @@
     <dependency>
       <groupId>org.kie.kogito</groupId>
       <artifactId>kogito-codegen-openapi</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.kie.kogito</groupId>
-      <artifactId>kogito-codegen-processes</artifactId>
     </dependency>
   </dependencies>
 </project>
