+++ kogito-quarkus-parent/kogito-quarkus-serverless-workflow-extension/kogito-quarkus-serverless-workflow-deployment/src/main/resources/META-INF/services/io.quarkus.deployment.dev.CompilationProvider
@@ -1,0 +1,1 @@
+org.kie.kogito.quarkus.serverless.workflow.deployment.ServerlessWorkflowCompilationProvider
